requirements.txt
# --- Interface e Web ---
flet==0.25.2
flet-desktop==0.25.2
flet-web==0.25.2
uvicorn==0.33.0
fastapi==0.124.4
reportlab

# --- Banco de Dados e Migrações ---
# psycopg2-binary é o driver para o PostgreSQL
psycopg2-binary
# SQLAlchemy facilitará as migrações futuras (Alembic)
sqlalchemy
# sqlite3 já vem no Python, não precisa listar

# --- Segurança e Utilidades ---
# Essencial para o hash de senhas que já implementamos
bcrypt
python-dotenv==1.0.1
pandas==2.0.3
openpyxl==3.1.5
pytest==8.3.5

# --- Dependências de Infraestrutura ---
anyio==4.5.2
httpx==0.28.1
pydantic==2.10.6
typing-extensions==4.13.2
